520 Move embedded-webhook-public-rest sources under public_ subpackage
Align the openapi generator's apiPackage/modelPackage with the module
name (webhook, not execution) and move the hand-written controllers,
converter, and test under the same com.bytechef.ee.embedded.webhook.public_.web.rest
namespace, matching the sibling embedded-execution-public-rest layout.
