+++ server/ee/libs/embedded/embedded-webhook/embedded-webhook-public-rest/build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
 }
 
 val generateOpenAPISpring by tasks.registering(org.openapitools.generator.gradle.plugin.tasks.GenerateTask::class) {
-    apiPackage.set("com.bytechef.ee.embedded.execution.public_.web.rest")
+    apiPackage.set("com.bytechef.ee.embedded.webhook.public_.web.rest")
     configOptions.set(
         mapOf(
             "interfaceOnly" to "true",
@@ -15,7 +15,7 @@ val generateOpenAPISpring by tasks.registering(org.openapitools.generator.gradle
     generatorName.set("spring")
     inputSpec.set( "$projectDir/openapi.yaml")
     modelNameSuffix.set("Model")
-    modelPackage.set("com.bytechef.ee.embedded.execution.public_.web.rest.model")
+    modelPackage.set("com.bytechef.ee.embedded.webhook.public_.web.rest.model")
     outputDir.set("$projectDir/generated")
     schemaMappings.set(
         mapOf(
