+++ server/ee/libs/embedded/embedded-webhook/embedded-webhook-public-rest/src/test/java/com/bytechef/ee/embedded/webhook/public_/web/rest/RequestTriggerApiControllerIntTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the Enterprise License.
  */
 
-package com.bytechef.ee.embedded.webhook;
+package com.bytechef.ee.embedded.webhook.public_.web.rest;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
